AE4350 Bio-Inspired Intelligence and Learning
All files necessary to run the continuous reinforcement learning DDPG agent for a 6 degree-of-freedom pole balancing environment.
Run only the main file. 
Once some iterations have been run, the run can be stopped and theanimation.m or learningPlots.m files can be used to produce an animation or learning plot for the iterations run. Only the iterations input must be changed to less than the number of iterations that were run in the main file.